scenario_landing_qwen/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12
jinja2==3.1.4
pyjwt==2.9.0
aiosqlite==0.20.0